requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.3
psycopg2>=2.8
redis>=3.5
celery>=5.0
channels-redis
requests